Added the thing removal
- things now can be deleted
- fixed alpha map checks
- fixed things that were never loaded

+++ SpaceExplorers/GuiCreation.cpp
@@ -2,12 +2,12 @@
 #include "Gui.h"
 
 
-ErrCode Gui::CreatePanel(Doh3d::GPanel*& pPanel, float pPosX, float pPosY, float pSizeX, float pSizeY, const std::string& pTextureName)
+ErrCode Gui::CreatePanel(Doh3d::GPanel*& pPanel, int pPosX, int pPosY, int pSizeX, int pSizeY, const std::string& pTextureName)
 {
 	LOG("Gui::CreatePanel()");
 	ErrCode3d err3d;
 
-	Doh3d::GPanel* panel = new Doh3d::GPanel(pPosX, pPosY, pSizeX, pSizeY, pTextureName);
+	Doh3d::GPanel* panel = new Doh3d::GPanel((float)pPosX, (float)pPosY, (float)pSizeX, (float)pSizeY, pTextureName);
 	
 	if (m_isLoaded)
 	{
@@ -25,13 +25,13 @@ ErrCode Gui::CreatePanel(Doh3d::GPanel*& pPanel, float pPosX, float pPosY, float
 }
 
 
-ErrCode Gui::CreateButton(Doh3d::GButton*& pButton, float pPosX, float pPosY, float pSizeX, float pSizeY,
+ErrCode Gui::CreateButton(Doh3d::GButton*& pButton, int pPosX, int pPosY, int pSizeX, int pSizeY,
 							const std::string& pTexNameNormal, const std::string& pTexNamePressed, const std::string& pTexNameSelected, const std::string& pTexNameDisabled)
 {
 	LOG("Gui::CreateButton()");
 	ErrCode3d err3d;
 
-	Doh3d::GButton* button = new Doh3d::GButton(pPosX, pPosY, pSizeX, pSizeY, pTexNameNormal, pTexNamePressed, pTexNameSelected, pTexNameDisabled);
+	Doh3d::GButton* button = new Doh3d::GButton((float)pPosX, (float)pPosY, (float)pSizeX, (float)pSizeY, pTexNameNormal, pTexNamePressed, pTexNameSelected, pTexNameDisabled);
 	
 	if (m_isLoaded)
 	{
@@ -48,12 +48,12 @@ ErrCode Gui::CreateButton(Doh3d::GButton*& pButton, float pPosX, float pPosY, fl
 	return err_noErr;
 }
 
-ErrCode Gui::CreateText(Doh3d::GText*& pGText, float pPosX, float pPosY, const std::string& pText, const std::string& pFontName)
+ErrCode Gui::CreateText(Doh3d::GText*& pGText, int pPosX, int pPosY, const std::string& pText, const std::string& pFontName)
 {
 	LOG("Gui::CreateText()");
 	ErrCode3d err3d;
 
-	Doh3d::GText* text = new Doh3d::GText(pPosX, pPosY, pText, pFontName);
+	Doh3d::GText* text = new Doh3d::GText((float)pPosX, (float)pPosY, pText, pFontName);
 
 	if (m_isLoaded)
 	{
@@ -70,13 +70,13 @@ ErrCode Gui::CreateText(Doh3d::GText*& pGText, float pPosX, float pPosY, const s
 	return err_noErr;
 }
 
-ErrCode Gui::CreateGGrid(Doh3d::GGrid*& pGGrid, float pPosX, float pPosY, float pSizeX, float pSizeY, int pNumberOfCells,
+ErrCode Gui::CreateGGrid(Doh3d::GGrid*& pGGrid, int pPosX, int pPosY, int pSizeX, int pSizeY, int pNumberOfCells,
 						   const std::string& pTextureName, const std::string& pFrameTextureName)
 {
 	LOG("Gui::CreateGGrid()");
 	ErrCode3d err3d;
 
-	Doh3d::GGrid* grid = new Doh3d::GGrid(pPosX, pPosY, pSizeX, pSizeY, pNumberOfCells, pTextureName, pFrameTextureName);
+	Doh3d::GGrid* grid = new Doh3d::GGrid((float)pPosX, (float)pPosY, (float)pSizeX, (float)pSizeY, pNumberOfCells, pTextureName, pFrameTextureName);
 
 	if (m_isLoaded)
 	{
